+++ packages/web/vite.config.js
@@ -10,7 +10,8 @@ export default defineConfig({
     },
   },
   server: {
-    port: 3000,
+    port: 3004,
+    strictPort: true,
     host: true,
   },
 })